bootstrap: delay searching for a compiler until we need it
